mv launch_engine, _set_envs_and_config
